// packages/shared/types/database.d.ts
export type Json = string | number | boolean | null | {
    [key: string]: Json | undefined;
} | Json[];
export type Database = {
    __InternalSupabase: {
        PostgrestVersion: "14.1";
    };
    public: {
        Tables: {
            accounts: {
                Row: {
                    account_type: string;
                    admin_id: string;
                    business_type: string | null;
                    country: string;
                    created_at: string;
                    currency: string | null;
                    data: Json;
                    deleted_at: string | null;
                    email: string | null;
                    id: string;
                    is_charges_enabled: boolean;
                    is_details_submitted: boolean;
                    is_payouts_enabled: boolean;
                    platform_fee_fixed: number | null;
                    platform_fee_percent: number | null;
                    processor_fees_applicable: boolean;
                    status: string;
                    stripe_id: string | null;
                    updated_at: string;
                };
                Insert: {
                    account_type?: string;
                    admin_id: string;
                    business_type?: string | null;
                    country: string;
                    created_at?: string;
                    currency?: string | null;
                    data?: Json;
                    deleted_at?: string | null;
                    email?: string | null;
                    id?: string;
                    is_charges_enabled?: boolean;
                    is_details_submitted?: boolean;
                    is_payouts_enabled?: boolean;
                    platform_fee_fixed?: number | null;
                    platform_fee_percent?: number | null;
                    processor_fees_applicable?: boolean;
                    status?: string;
                    stripe_id?: string | null;
                    updated_at?: string;
                };
                Update: {
                    account_type?: string;
                    admin_id?: string;
                    business_type?: string | null;
                    country?: string;
                    created_at?: string;
                    currency?: string | null;
                    data?: Json;
                    deleted_at?: string | null;
                    email?: string | null;
                    id?: string;
                    is_charges_enabled?: boolean;
                    is_details_submitted?: boolean;
                    is_payouts_enabled?: boolean;
                    platform_fee_fixed?: number | null;
                    platform_fee_percent?: number | null;
                    processor_fees_applicable?: boolean;
                    status?: string;
                    stripe_id?: string | null;
                    updated_at?: string;
                };
                Relationships: [
                    {
                        foreignKeyName: "accounts_admin_id_fkey";
                        columns: ["admin_id"];
                        isOneToOne: false;
                        referencedRelation: "users";
                        referencedColumns: ["id"];
                    }
                ];
            };
            organizations: {
                Row: {
                    account_id: string | null;
                    avatar_url: string | null;
                    blocked_at: string | null;
                    created_at: string;
                    customer_email_settings: Json;
                    customer_invoice_next_number: number;
                    customer_invoice_prefix: string;
                    customer_portal_settings: Json;
                    deleted_at: string | null;
                    details: Json;
                    details_submitted_at: string | null;
                    email: string | null;
                    feature_settings: Json;
                    id: string;
                    name: string;
                    notification_settings: Json;
                    onboarded_at: string | null;
                    order_settings: Json;
                    profile_settings: Json;
                    slug: string;
                    socials: Json;
                    status: string;
                    status_updated_at: string | null;
                    subscription_settings: Json;
                    subscriptions_billing_engine: boolean;
                    updated_at: string;
                    website: string | null;
                };
                Insert: {
                    account_id?: string | null;
                    avatar_url?: string | null;
                    blocked_at?: string | null;
                    created_at?: string;
                    customer_email_settings?: Json;
                    customer_invoice_next_number?: number;
                    customer_invoice_prefix?: string;
                    customer_portal_settings?: Json;
                    deleted_at?: string | null;
                    details?: Json;
                    details_submitted_at?: string | null;
                    email?: string | null;
                    feature_settings?: Json;
                    id?: string;
                    name: string;
                    notification_settings?: Json;
                    onboarded_at?: string | null;
                    order_settings?: Json;
                    profile_settings?: Json;
                    slug: string;
                    socials?: Json;
                    status?: string;
                    status_updated_at?: string | null;
                    subscription_settings?: Json;
                    subscriptions_billing_engine?: boolean;
                    updated_at?: string;
                    website?: string | null;
                };
                Update: {
                    account_id?: string | null;
                    avatar_url?: string | null;
                    blocked_at?: string | null;
                    created_at?: string;
                    customer_email_settings?: Json;
                    customer_invoice_next_number?: number;
                    customer_invoice_prefix?: string;
                    customer_portal_settings?: Json;
                    deleted_at?: string | null;
                    details?: Json;
                    details_submitted_at?: string | null;
                    email?: string | null;
                    feature_settings?: Json;
                    id?: string;
                    name?: string;
                    notification_settings?: Json;
                    onboarded_at?: string | null;
                    order_settings?: Json;
                    profile_settings?: Json;
                    slug?: string;
                    socials?: Json;
                    status?: string;
                    status_updated_at?: string | null;
                    subscription_settings?: Json;
                    subscriptions_billing_engine?: boolean;
                    updated_at?: string;
                    website?: string | null;
                };
                Relationships: [
                    {
                        foreignKeyName: "fk_organizations_account_id";
                        columns: ["account_id"];
                        isOneToOne: false;
                        referencedRelation: "accounts";
                        referencedColumns: ["id"];
                    }
                ];
            };
// ...
        };
// ...
    };
};